View: Fix timeline scrollbar drag being overwritten by fractional scroll offsets
ImGui stores window scroll as whole pixels, so a sub-pixel remainder is not a real pending wheel request. Treat deltas below one pixel as already applied, keep m_previous_scroll_position as a float, and prevent the scrollbar from fighting with accumulated wheel/hotkey offsets. Update the UI guidance to document the rounding behavior.

+++ src/view/src/rocprofvis_timeline_view.h
@@ -219,7 +219,7 @@ private:
     float                               m_scroll_position_y;
     float                               m_content_max_y_scroll;
     bool                                m_can_drag_to_pan;
-    double                              m_previous_scroll_position;
+    float                               m_previous_scroll_position;
     bool                                m_meta_map_made;
     bool                                m_resize_activity;
     bool                                m_reorder_auto_scrolling;
